requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
-e git+https://github.com/freedomofpress/django-logging.git@dfd8904#egg=django-logging-json
beautifulsoup4==4.6.0     # via wagtail
certifi==2017.4.17        # via requests
chardet==3.0.4            # via requests
click==6.7                # via pip-tools
django-modelcluster==3.1
django-settings-export==1.2.1
django-taggit==0.22.1     # via wagtail
django-treebeard==4.1.2   # via wagtail
django-webpack-loader==0.5.0
django==1.11.3
djangorestframework==3.6.3  # via wagtail
elasticsearch==2.4.1
factory-boy==2.8.1
faker==0.7.17             # via factory-boy
first==2.0.1              # via pip-tools
gunicorn==19.7.1
html5lib==0.999999999     # via wagtail
idna==2.5                 # via requests
ipaddress==1.0.18         # via faker
olefile==0.44             # via pillow
pillow==4.2.1             # via wagtail
pip-tools==1.9.0
psycopg2==2.7.1
python-dateutil==2.6.1    # via faker
pytz==2017.2              # via django, django-modelcluster
requests==2.18.1          # via wagtail
six==1.10.0               # via faker, html5lib, pip-tools, python-dateutil, unittest-xml-reporting
unidecode==0.4.21         # via wagtail
unittest-xml-reporting==2.1.0
urllib3==1.21.1           # via elasticsearch, requests
wagtail-factories==0.3.0
wagtail-metadata==0.3.1
wagtail==1.11.1
webencodings==0.5.1       # via html5lib
willow==0.4               # via wagtail
